[back] migrate user to "sql" entity (doesn't compile)

+++ src/main/java/fr/sii/controller/common/user/CommonProfilImageController.java
@@ -7,8 +7,8 @@ import com.google.appengine.api.blobstore.BlobstoreServiceFactory;
 import fr.sii.config.global.GlobalSettings;
 import fr.sii.domain.common.Uri;
 import fr.sii.domain.exception.NotFoundException;
-import fr.sii.domain.user.User;
-import fr.sii.domain.user.UserProfil;
+import fr.sii.entity.User;
+import fr.sii.dto.user.UserProfil;
 import fr.sii.service.user.UserService;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -18,11 +18,6 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-
-/**
- * Created by tmaugin on 09/07/2015.
- * SII
- */
 
 @RequestMapping(value="/api/profil")
 public class CommonProfilImageController {
@@ -50,7 +45,7 @@ public class CommonProfilImageController {
      */
     @RequestMapping(value="/image/{blobKeyParam}", method= RequestMethod.GET)
     @ResponseBody
-    public void getProfileImage(HttpServletRequest req, HttpServletResponse res,  @PathVariable String blobKeyParam) throws IOException {
+    public void getProfileImage(HttpServletRequest req, HttpServletResponse res, @PathVariable String blobKeyParam) throws IOException {
         BlobKey blobKey = new BlobKey(blobKeyParam);
         blobstoreService.serve(blobKey, res);
     }
@@ -65,7 +60,7 @@ public class CommonProfilImageController {
      */
     @RequestMapping(value="/image/user/{id}", method= RequestMethod.GET)
     @ResponseBody
-    public void getProfileImageByUserId(HttpServletRequest req, HttpServletResponse res,  @PathVariable Long id) throws IOException, NotFoundException {
+    public void getProfileImageByUserId(HttpServletRequest req, HttpServletResponse res,  @PathVariable Integer id) throws IOException, NotFoundException {
         User u = userService.findById(id);
         if(u == null)
         {
@@ -96,7 +91,7 @@ public class CommonProfilImageController {
      */
     @RequestMapping(value="/image/user/url/{id}", method= RequestMethod.GET)
     @ResponseBody
-    public Uri getProfilImageUrlByUserId(HttpServletRequest req, HttpServletResponse res,  @PathVariable Long id) throws IOException, NotFoundException {
+    public Uri getProfilImageUrlByUserId(HttpServletRequest req, HttpServletResponse res, @PathVariable Integer id) throws IOException, NotFoundException {
         User u = userService.findById(id);
         if(u == null)
         {
